Update to correct docker package.
This was caused by some real life testing that resulted in the realization
that the incorrect package was being used for docker.

These changes resulted in some test errors related to old vs new methods
and the way they are accessed. These tests have currently been skipped
and will be updated soon