Add initial queue and fix tests
Pulling apart the queue, db, and runner so they can be composed/tested more easily

This includes initial experiments with doing cleanup in the background and both
telling the runner when to run more tests via a return on complete, and also
during signal phases when there are more items to run based on cleanup analysis.

More work will need to be done on these aspects in the future.